[feature] #13 쿠폰 서비스 수정, 일부 리팩토링

+++ src/main/java/com/fouribnb/coupon/presentation/dto/request/UpdateCouponRequestDto.java
@@ -9,6 +9,6 @@ import lombok.Getter;
 public class UpdateCouponRequestDto {
 
     private String couponName;
-    private int discountValue;
+    private Long discountValue;
 
 }
