+++ src/main/java/com/fouribnb/coupon/presentation/dto/response/UpdateCouponResponseDto.java
@@ -13,7 +13,7 @@ public class UpdateCouponResponseDto {
     private UUID paymentId;
     private String couponName;
     private CouponStatus couponStatus;
-    private int discountValue;
+    private Long discountValue;
     private Boolean isUsed;
 
 }
